update format date number and logo
