+++ src/styles/inputStyles.scss
@@ -19,6 +19,8 @@ Contact: wandaazhar@gmail.com
 
     label {
       font-size: 1.4rem;
+      color: grey;
+      color: grey;
     }
 
     .formControl {
@@ -47,6 +49,7 @@ Contact: wandaazhar@gmail.com
 
     label {
       font-size: 1.4rem;
+      color: grey;
     }
 
     &.input1 {
@@ -103,6 +106,7 @@ Contact: wandaazhar@gmail.com
 
     label {
       font-size: 1.4rem;
+      color: grey;
     }
 
     &.input1 {
@@ -135,6 +139,7 @@ Contact: wandaazhar@gmail.com
 
     label {
       font-size: 1.4rem;
+      color: grey;
     }
 
     &.input1 {
