+++ src/styles/variable.scss
@@ -7,7 +7,7 @@ Contact: wandaazhar@gmail.com
 // BACKGROUND
 $mainBackground: linear-gradient(217deg, rgba(238, 174, 202, 1) 0%, rgba(148, 187, 233, 1) 100%);
 $stepBackground: linear-gradient(90deg, rgb(0, 90, 245) 0%, rgb(75, 150, 235) 100%);
-$backgroundBlue: #0071ce;
+$backgroundBtn: #eb8912;
 $colorBackgroundDark: #263237;
 $colorBackgroundNavbar: #314148;
 $colorTeal: #26A79A;
